Extract Layer 2 match helpers
